requirements.txt
appdirs==1.4.4
attrs==20.3.0
autopep8==1.5.6
certifi==2020.12.5
click==7.1.2
coverage==5.5
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Caching==1.10.1
greenlet==1.0.0
gunicorn==20.1.0
importlib-metadata==4.0.0
importlib-resources==5.1.2
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pipenv==2020.11.15
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-flask==1.2.0
python-dotenv==0.17.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.4.9
toml==0.10.2
typing-extensions==3.7.4.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.4.1
